+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.0.7] - 2016-02-05
+### Changed
+- `rosalind::fib` return type now `Result`, all modules now return `Result`
+
+## [0.0.6] - 2016-02-04
+### Added
+- `rosalind::hamm` module
+
 ## [0.0.5] - 2016-01-25
 ### Added
 - `rosalind::prot` module